requirements.txt
Django==4.1.10
python-dotenv==1.0.0
djangorestframework==3.13.1
Pillow==10.0.0
django-phonenumber-field==7.1.0
phonenumbers==8.13.17
drf-yasg2==1.19.4
django-filter==23.2
django-import-export==3.2.0